+++ packages/block-library/src/embed/util.js
@@ -1,6 +1,5 @@
 import clsx from 'clsx';
 import memoize from 'memize';
-import { privateApis as componentsPrivateApis } from '@wordpress/components';
 import { renderToString } from '@wordpress/element';
 import {
 	createBlock,
@@ -8,12 +7,11 @@ import {
 	getBlockVariations,
 } from '@wordpress/blocks';
 import { getAuthority } from '@wordpress/url';
+import { kebabCase } from '@wordpress/kebab-case';
 import metadata from './block.json';
 import { ASPECT_RATIOS, WP_EMBED_TYPE } from './constants';
-import { unlock } from '../lock-unlock';
 
 const { name: DEFAULT_EMBED_BLOCK } = metadata;
-const { kebabCase } = unlock( componentsPrivateApis );
 
 /** @typedef {import('@wordpress/blocks').WPBlockVariation} WPBlockVariation */